Value stuff isn't output-specific
